backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
scikit-learn
pandas
numpy
joblib
python-dotenv
